fix: remove old commitLines
