+++ packages/ui/src/stories/commitLines/DemoCommitLines.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import LineGroup from '$lib/commitLines/LineGroup.svelte';
+	import Line from '$lib/commitLines/Line.svelte';
 	import { LineManager } from '$lib/commitLines/lineManager';
 	import type { CommitData } from '$lib/commitLines/types';
 
@@ -8,53 +8,39 @@
 		localCommits: CommitData[];
 		localAndRemoteCommits: CommitData[];
 		integratedCommits: CommitData[];
-		sameForkpoint: boolean;
 	}
 
-	const {
-		remoteCommits,
-		localCommits,
-		localAndRemoteCommits,
-		integratedCommits,
-		sameForkpoint
-	}: Props = $props();
+	const { remoteCommits, localCommits, localAndRemoteCommits, integratedCommits }: Props = $props();
 
 	const lineManager = $derived(
-		new LineManager(
-			{ remoteCommits, localCommits, localAndRemoteCommits, integratedCommits },
-			sameForkpoint
-		)
+		new LineManager({ remoteCommits, localCommits, localAndRemoteCommits, integratedCommits })
 	);
 </script>
 
 <div class="column">
 	{#each remoteCommits as commit}
 		<div class="group">
-			<LineGroup lineGroup={lineManager.get(commit.id)} />
+			<Line line={lineManager.get(commit.id)} />
 		</div>
 	{/each}
 
 	{#each localCommits as commit}
 		<div class="group">
-			<LineGroup lineGroup={lineManager.get(commit.id)} />
+			<Line line={lineManager.get(commit.id)} />
 		</div>
 	{/each}
 
 	{#each localAndRemoteCommits as commit}
 		<div class="group">
-			<LineGroup lineGroup={lineManager.get(commit.id)} />
+			<Line line={lineManager.get(commit.id)} />
 		</div>
 	{/each}
 
 	{#each integratedCommits as commit}
 		<div class="group">
-			<LineGroup lineGroup={lineManager.get(commit.id)} />
+			<Line line={lineManager.get(commit.id)} />
 		</div>
 	{/each}
-
-	<div class="group">
-		<LineGroup lineGroup={lineManager.base} />
-	</div>
 </div>
 
 <style lang="postcss">
